Refactor host routing and add scoped maintenance windows
