+++ server/app/routers/__init__.py
@@ -1,1 +1,1 @@
-from . import agent, ansible, approvals, audit, auth, backup_verification, hosts, jobs, migrations, mfa, search, terminal_ws, ui  # noqa: F401
+from . import agent, ansible, approvals, audit, auth, backup_verification, host_observability, host_packages, host_services, host_users, hosts, jobs, maintenance_windows, migrations, mfa, search, terminal_ws, ui  # noqa: F401
