+++ server/tests/test_backup_verification_api.py
@@ -1,24 +1,10 @@
-import importlib
 import sqlite3
 
-
-def _boot_app(monkeypatch):
-    monkeypatch.setenv("DATABASE_URL", "sqlite+pysqlite:///:memory:")
-    monkeypatch.setenv("BOOTSTRAP_USERNAME", "admin")
-    monkeypatch.setenv("BOOTSTRAP_PASSWORD", "admin-password-123")
-    monkeypatch.setenv("UI_COOKIE_SECURE", "false")
-    monkeypatch.setenv("ALLOW_INSECURE_NO_AGENT_TOKEN", "true")
-    monkeypatch.setenv("AGENT_SHARED_TOKEN", "")
-    monkeypatch.setenv("DB_AUTO_CREATE_TABLES", "true")
-    monkeypatch.setenv("DB_REQUIRE_MIGRATIONS_UP_TO_DATE", "false")
-    monkeypatch.setenv("MFA_REQUIRE_FOR_PRIVILEGED", "false")
-
-    app_factory = importlib.import_module("app.app_factory")
-    return app_factory.create_app()
+from conftest import bootstrap_test_app
 
 
-def test_backup_verification_happy_path(monkeypatch, tmp_path):
-    app = _boot_app(monkeypatch)
+def test_backup_verification_happy_path(monkeypatch, tmp_path, auth_client_factory):
+    app = bootstrap_test_app(monkeypatch)
 
     db_path = tmp_path / "backup.sqlite"
     conn = sqlite3.connect(str(db_path))
@@ -29,14 +15,7 @@ def test_backup_verification_happy_path(monkeypatch, tmp_path):
     conn.commit()
     conn.close()
 
-    from fastapi.testclient import TestClient
-
-    with TestClient(app) as client:
-        r = client.post("/auth/login", json={"username": "admin", "password": "admin-password-123"})
-        assert r.status_code == 200, r.text
-        csrf = client.cookies.get("fleet_csrf")
-        headers = {"X-CSRF-Token": csrf} if csrf else {}
-
+    with auth_client_factory(app) as (client, headers):
         run = client.post(
             "/backup-verification/runs",
             json={
@@ -58,8 +37,8 @@ def test_backup_verification_happy_path(monkeypatch, tmp_path):
         assert latest.json()["id"] == data["id"]
 
 
-def test_backup_verification_schema_mismatch_fails(monkeypatch, tmp_path):
-    app = _boot_app(monkeypatch)
+def test_backup_verification_schema_mismatch_fails(monkeypatch, tmp_path, auth_client_factory):
+    app = bootstrap_test_app(monkeypatch)
 
     db_path = tmp_path / "backup-low-schema.sqlite"
     conn = sqlite3.connect(str(db_path))
@@ -69,14 +48,7 @@ def test_backup_verification_schema_mismatch_fails(monkeypatch, tmp_path):
     conn.commit()
     conn.close()
 
-    from fastapi.testclient import TestClient
-
-    with TestClient(app) as client:
-        r = client.post("/auth/login", json={"username": "admin", "password": "admin-password-123"})
-        assert r.status_code == 200, r.text
-        csrf = client.cookies.get("fleet_csrf")
-        headers = {"X-CSRF-Token": csrf} if csrf else {}
-
+    with auth_client_factory(app) as (client, headers):
         run = client.post(
             "/backup-verification/runs",
             json={
