+++ server/tests/test_backup_verification_policy_api.py
@@ -1,25 +1,11 @@
-import importlib
 import sqlite3
 from datetime import datetime, timedelta, timezone
 
-
-def _boot_app(monkeypatch):
-    monkeypatch.setenv("DATABASE_URL", "sqlite+pysqlite:///:memory:")
-    monkeypatch.setenv("BOOTSTRAP_USERNAME", "admin")
-    monkeypatch.setenv("BOOTSTRAP_PASSWORD", "admin-password-123")
-    monkeypatch.setenv("UI_COOKIE_SECURE", "false")
-    monkeypatch.setenv("ALLOW_INSECURE_NO_AGENT_TOKEN", "true")
-    monkeypatch.setenv("AGENT_SHARED_TOKEN", "")
-    monkeypatch.setenv("DB_AUTO_CREATE_TABLES", "true")
-    monkeypatch.setenv("DB_REQUIRE_MIGRATIONS_UP_TO_DATE", "false")
-    monkeypatch.setenv("MFA_REQUIRE_FOR_PRIVILEGED", "false")
-
-    app_factory = importlib.import_module("app.app_factory")
-    return app_factory.create_app()
+from conftest import bootstrap_test_app
 
 
-def test_backup_verification_policy_run_now_and_notifications(monkeypatch, tmp_path):
-    app = _boot_app(monkeypatch)
+def test_backup_verification_policy_run_now_and_notifications(monkeypatch, tmp_path, auth_client_factory):
+    app = bootstrap_test_app(monkeypatch)
 
     db_path = tmp_path / "backup.sqlite"
     conn = sqlite3.connect(str(db_path))
@@ -29,14 +15,7 @@ def test_backup_verification_policy_run_now_and_notifications(monkeypatch, tmp_p
     conn.commit()
     conn.close()
 
-    from fastapi.testclient import TestClient
-
-    with TestClient(app) as client:
-        r = client.post("/auth/login", json={"username": "admin", "password": "admin-password-123"})
-        assert r.status_code == 200, r.text
-        csrf = client.cookies.get("fleet_csrf")
-        headers = {"X-CSRF-Token": csrf} if csrf else {}
-
+    with auth_client_factory(app) as (client, headers):
         put = client.put(
             "/backup-verification/policy",
             json={
