+++ server/tests/test_phase3_host_filters_frontend_smoke.py
@@ -1,23 +1,16 @@
-from pathlib import Path
-
-
-def test_phase3_host_filters_scripts_are_split_and_loaded():
-    root = Path(__file__).resolve().parents[2]
-    templates = root / "server" / "app" / "templates"
-
-    index_html = (templates / "index.html").read_text(encoding="utf-8")
+def test_phase3_host_filters_scripts_are_split_and_loaded(templates_dir):
+    index_html = (templates_dir / "index.html").read_text(encoding="utf-8")
     assert '/assets/fleet-phase3-host-filters-ui.js' in index_html
     assert '/assets/fleet-phase3-host-filters-vuln.js' in index_html
     assert '/assets/fleet-phase3-host-filters.js' in index_html
 
-    orchestrator = (templates / "fleet-phase3-host-filters.js").read_text(encoding="utf-8")
+    orchestrator = (templates_dir / "fleet-phase3-host-filters.js").read_text(encoding="utf-8")
     assert 'phase3HostFiltersUi.initHostFiltersUi' in orchestrator
     assert 'phase3HostFiltersVuln.initHostFiltersVuln' in orchestrator
 
 
-def test_phase3_vuln_upgrade_status_copy_smoke():
-    root = Path(__file__).resolve().parents[2]
-    vuln_js = (root / "server" / "app" / "templates" / "fleet-phase3-host-filters-vuln.js").read_text(encoding="utf-8")
+def test_phase3_vuln_upgrade_status_copy_smoke(templates_dir):
+    vuln_js = (templates_dir / "fleet-phase3-host-filters-vuln.js").read_text(encoding="utf-8")
 
     # CVE flow status lines used by UI and support docs/operators.
     assert 'Running CVE check…' in vuln_js
